Remove unused CLI bulk wait support

Copilot-Session: 14e6fead-9dad-4880-bdaf-75663b728b67
