+++ src/Aspire.Cli/Aspire.Cli.csproj
@@ -158,6 +158,7 @@
     <Compile Include="$(SharedDir)IConfigurationExtensions.cs" Link="Utils\IConfigurationExtensions.cs" />
     <Compile Include="$(RepoRoot)src\Aspire.Hosting\Backchannel\BackchannelDataTypes.cs" Link="Backchannel\CliBackchannelDataTypes.cs" />
     <Compile Include="$(RepoRoot)src\Aspire.Hosting\ApplicationModel\KnownResourceCommands.cs" Link="Utils\KnownResourceCommands.cs" />
+    <Compile Include="$(RepoRoot)src\Aspire.Hosting\ApplicationModel\KnownResourceStates.cs" Link="Utils\KnownResourceStates.cs" />
     <Compile Include="$(SharedDir)PackageUpdateHelpers.cs" Link="Utils\PackageUpdateHelpers.cs" />
     <Compile Include="$(SharedDir)PathLookupHelper.cs" Link="Utils\PathLookupHelper.cs" />
     <Compile Include="$(SharedDir)BackchannelConstants.cs" Link="Backchannel\BackchannelConstants.cs" />
